requirements.txt
oauthlib==3.1.1
google_auth_oauthlib==0.4.4
discord.py==1.7.3
discord==1.7.3
google_api_python_client==2.12.0
protobuf==3.17.3
python-dotenv==0.18.0
python_dateutil==2.8.1
